requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
billiard==4.2.1
celery==5.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-ipware==7.0.1
django-ratelimit==4.1.0
ipgeolocationio==1.0.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
redis==5.0.8
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
